fix: small bugs in data set
Discovered 2 small bugs in the way the intial data sets were formed 1) ~80 duplicate rows and 2) for the sensitivities, the centering of temp and sal was done slightly differently. These are fixed and now the code has be rerun for the model results, as well as figures. Some additional tables may need to be corrected.

Note that I verified that the datasets from the sensitivities now match the base model for the scenarios it should and that there are no duplicate rows in the dataset.